+++ packages/electrode-react-webapp/lib/hapi/index.js
@@ -9,32 +9,40 @@ const HTTP_REDIRECT = 302;
 
 const registerRoutes = (server, options, next) => {
   ReactWebapp.setupOptions(options)
-    .then((registerOptions) => {
+    .then(registerOptions => {
       _.each(registerOptions.paths, (v, path) => {
-        assert(v.content, `You must define content for the webapp plugin path ${path}`);
-        const routeHandler = ReactWebapp.makeRouteHandler(
-          registerOptions, ReactWebapp.resolveContent(v.content)
-        );
+        const resolveContent = () => {
+          if (registerOptions.serverSideRendering !== false) {
+            assert(
+              v.content !== undefined,
+              `You must define content for the webapp plugin path ${path}`
+            );
+            return ReactWebapp.resolveContent(v.content);
+          }
+          return "";
+        };
+
+        const routeHandler = ReactWebapp.makeRouteHandler(registerOptions, resolveContent());
 
         server.route({
           method: v.method || "GET",
           path,
           config: v.config || {},
           handler: (request, reply) => {
-            const handleStatus = (data) => {
+            const handleStatus = data => {
               const status = data.status;
               if (status === HTTP_REDIRECT) {
                 reply.redirect(data.path);
               } else {
-                reply({message: "error"}).code(status);
+                reply({ message: "error" }).code(status);
               }
             };
 
-            routeHandler({mode: request.query.__mode || "", request})
-              .then((data) => {
+            routeHandler({ mode: request.query.__mode || "", request })
+              .then(data => {
                 return data.status ? handleStatus(data) : reply(data);
               })
-              .catch((err) => {
+              .catch(err => {
                 reply(err.message).code(err.status || HTTP_ERROR_500);
               });
           }